Move react-query fetches from issues components to separate files
