+++ features/issue/components/index.ts
@@ -1,1 +1,2 @@
 export * from './NewIssueForm'
+export * from './IssuesList'
