+++ shared/api/issues/deleteIssue.ts
@@ -1,0 +1,13 @@
+const API_PATH = process.env.NEXT_PUBLIC_API_PATH
+
+export const deleteIssue = async (id: number) => {
+	const response = await fetch(`/api/issues`, {
+		method: 'DELETE',
+		headers: {
+			'Content-Type': `${API_PATH}/api/issues`,
+		},
+		body: JSON.stringify({ id }),
+	})
+	const issue = await response.json()
+	return issue
+}
