+++ shared/api/issues/index.ts
@@ -1,0 +1,4 @@
+export * from './deleteIssue'
+export * from './getIssues'
+export * from './useDeleteIssue'
+export * from './useGetIssues'
