+++ shared/api/issues/useDeleteIssue.ts
@@ -1,0 +1,27 @@
+import { useMutation, useQueryClient } from '@tanstack/react-query'
+import { deleteIssue } from '@/shared/api/issues'
+import { useToast } from '@/components/ui/use-toast'
+
+export const useDeleteIssue = () => {
+	const { toast } = useToast()
+	const queryClient = useQueryClient()
+
+	const deleteIssueMutation = useMutation({
+		mutationFn: deleteIssue,
+		onSuccess: () => {
+			queryClient.invalidateQueries({ queryKey: ['issues'] })
+			toast({
+				title: 'Success!',
+				description: 'Deleting issue succeed.',
+			})
+		},
+		onError: () => {
+			toast({
+				title: 'Error',
+				description: 'Something went wrong',
+				variant: 'destructive',
+			})
+		},
+	})
+	return deleteIssueMutation
+}
