+++ shared/api/issues/useGetIssues.ts
@@ -1,0 +1,10 @@
+import { useQuery } from '@tanstack/react-query'
+import { getIssues } from '@/shared/api/issues'
+
+export const useGetIssues = () => {
+	const getIssuesQuery = useQuery({
+		queryKey: ['issues'],
+		queryFn: getIssues,
+	})
+	return getIssuesQuery
+}
